deleted polyfill and added polindrome
